Fixes for MaxHydration turned to const.
* Some extra fixes for the MaxHydrationStorage made to constant. The patches related to it were also moved to another file to not clutter the main patches file.
